+++ src/components/faq/Faq.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Faq = () => (
+  <div>
+    <p>Faqs</p>
+  </div>
+);
+
+export default Faq;
