+++ src/pages/contact.js
@@ -1,38 +1,11 @@
 import { NextSeo } from 'next-seo';
-import swal from 'sweetalert';
 import config from '../utils/config';
 
 import contactData from '../../content/general/contact.yaml';
-import ContactForm from '../components/contact/ContactForm';
 import Layout from '../components/Layout';
+import ContactInfo from '../components/contact/ContactInfo';
 
 function ContactPage() {
-  const handleSubmit = async (e) => {
-    // eslint-disable-next-line no-undef
-    const res = await fetch('/api/subscribe', {
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      method: 'POST',
-      body: JSON.stringify({
-        email: e.email,
-        firstName: e.firstName,
-        lastName: e.lastName,
-        message: e.message,
-        phone: e.phone,
-        company: e.company,
-        projectType: e.projectType,
-      }),
-    });
-
-    const { error } = await res.json();
-    console.log(error, 'error');
-    if (error === 'Success') {
-      swal('Success!', '🎉  Subscribed successfully', 'success');
-    } else {
-      swal('Already Subscribed!', error, 'error');
-    }
-  };
   return (
     <Layout>
       <NextSeo
@@ -53,66 +26,7 @@ function ContactPage() {
           ],
         }}
       />
-      <section className="section contact-us bg-white  is-hero-fade">
-        <div className="container">
-          <div className="row justify-content-center">
-            <div className="col-xl-11">
-              <div className="block">
-                <div className="row justify-content-center">
-                  <div className="col-lg-5 mb-5 mb-lg-0">
-                    <div className="pr-0 pr-lg-4">
-                      <h4 className="text-dark mb-3">Contact Us</h4>
-                      <p>
-                        Share the details of your project – like scope, time
-                        frames, or business challenges. Our team will carefully
-                        study them and then we’ll figure out the next move
-                        together.
-                      </p>
-
-                      <div className="mt-4 pt-2">
-                        <p className="text-uppercase small font-weight-500 mb-1">
-                          Email us:
-                        </p>
-                        <p>
-                          <a
-                            className="line-link text-dark"
-                            href={`mailto:${contactData.emailSales}`}
-                          >
-                            {contactData.emailSales}
-                          </a>
-                        </p>
-                      </div>
-                      <div className="mt-4">
-                        <p className="text-uppercase small font-weight-500 mb-1">
-                          Call us:
-                        </p>
-                        <p>
-                          <a
-                            className="line-link text-dark"
-                            href={`tel:${contactData.telephoneUk}`}
-                          >
-                            {contactData.telephoneUk}
-                          </a>
-                        </p>
-                      </div>
-                      <div className="mt-4">
-                        <p className="text-uppercase small font-weight-500 mb-1">
-                          Our Offices:
-                        </p>
-                        <p>{contactData.locationUk}</p>
-                        <p>{contactData.locationIndia}</p>
-                      </div>
-                    </div>
-                  </div>
-                  <div className="col-lg-7">
-                    <ContactForm onSubmit={(val) => handleSubmit(val)} />
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
+      <ContactInfo contactData={contactData} />
     </Layout>
   );
 }
